+++ src/main/BoxStyles.ts
@@ -90,6 +90,12 @@ const VERTICALBAR_PLUS_MINUS = new BoxStyle(
   '+',
   ['+', '+', '+', '+']
 );
+const VERTICALBAR_PLUS_MINUS_SLASHES = new BoxStyle(
+  ['-', '-', '|', '|'],
+  ['/', '\\', '\\', '/'],
+  '+',
+  ['+', '+', '+', '+']
+);
 
 export const BoxStyles = {
   THIN,
@@ -107,4 +113,5 @@ export const BoxStyles = {
   INNER_DOUBLE_OUTER_THIN_CURVE,
   INNER_THICK_OUTER_THIN_CURVE,
   VERTICALBAR_PLUS_MINUS,
+  VERTICALBAR_PLUS_MINUS_SLASHES,
 };